+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## [0.0.1] - TODO: Add release date.
 
 * TODO: GleamTextField 增加输入验证规则
+* TODO: GleamTextField 增加显示字数统计
+* TODO: GleamTextField 增加clearable 是否启用清除图标，点击清除图标后会清空输入框